% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_functions.R
\name{spike_slab_logistic}
\alias{spike_slab_logistic}
\title{spike_slab_logistic}
\usage{
spike_slab_logistic(
  chain_length,
  X,
  y,
  tau0,
  tau1,
  q,
  rinit = NULL,
  verbose = FALSE,
  burnin = 0,
  store = TRUE,
  Xt = NULL,
  XXt = NULL
)
}
\arguments{
\item{chain_length}{Markov chain length}

\item{X}{matrix of length n by p}

\item{y}{Response}

\item{tau0}{prior hyperparameter (non-negative real)}

\item{tau1}{prior hyperparameter (non-negative real)}

\item{q}{prior hyperparameter (strictly between 0 and 1)}

\item{X_transpose}{Pre-calculated transpose of X}
}
\value{
Output from Markov chain targeting the posterior corresponding to
Bayesian logistic regression with spike and slab priors
}
\description{
Generates Markov chain targeting the posterior corresponding to
Bayesian logistic regression with spike and slab priors
}
