% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{spike_slab_params}
\alias{spike_slab_params}
\title{spike_slab_params}
\usage{
spike_slab_params(n, p)
}
\arguments{
\item{n}{number of observations}

\item{p}{number of covariates}
}
\value{
spike-and-slab hyperparameters q, tau0, tau1, a0, b0
}
\description{
Generates hyperparameters for spike-and-slab
}
\examples{
hyper_params <- spike_slab_params(n=100,p=200)
print(hyper_params)
}
